B-19 Add User deposits manage form

+++ BankApplication/BankApplication/Constants.cs
@@ -4,6 +4,6 @@ namespace BankApplication {
     public static class Constants {
 
         //Кількість днів в році за якою буде розраховуватись відсоток.
-        public static readonly int K = 365;
+        public static readonly double K = 365;
     }
 }
